CLI-161: Fix #508: Install Drush 9 aliases to project dir
